requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
opencv-python>=4.5.5
tqdm>=4.64.0
scikit-image>=0.19.0
real-esrgan
gdown  # for downloading pretrained weights
requests
Flask==2.3.3
Werkzeug==2.3.7  
scipy               


